pyproject.toml
[tool.poetry]
name = "baserun-demo"
version = "0.1.0"
description = ""
authors = ["Erik Peterson <e@eriklp.com>"]
packages = [{ include = "agent", from = "." }, { include = "chatbot", from = "." }]

readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.12"
langchain = "^0.0.341"
python-dotenv = "^1.0.0"
openai = "^1.2.2"
wikipedia = "^1.4.0"
google-search-results = "^2.4.2"
anthropic = "^0.3.11"
rapidfuzz = "^3.5.1"
python-levenshtein = "^0.23.0"
wolframalpha = "^5.0.0"
baserun = { path = "/Users/epeterson/src/baserun/baserun-py", develop = true }

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
